allow omitting 'Log checksum' string in foreign logs
For EAC logs, it's expected that for logs with checksum the string
will be '==== Log checksum ... ===='. For foreign logs, it's also
assumed that it'll be like that, but given the nature of translation
files, it's possible that 'Log checksum' wasn't translated and omitted
and thus doesn't appear there, however the checksum is still valid.